pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["asyncpg>=0.30.0", "lightrag-hku>=1.3.8", "ollama>=0.5.1"]
